# Conflicts:
#	src/main/resources/application.properties

+++ src/main/java/ua/smartsub/smartsub/dao/RoleDao.java
@@ -2,9 +2,9 @@ package ua.smartsub.smartsub.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import ua.smartsub.smartsub.entity.Role;
+import ua.smartsub.smartsub.model.entity.Role;
 
-public interface RoleDao  extends JpaRepository<Role,Integer> {
+public interface RoleDao  extends JpaRepository<Role,Long> {
 
     @Query("select r from Role r where r.name=:name")
     Role findByName(String name);
